# flate: improve huffman flate hcode spatial locality