Improving huffman `hcode` spatial locality and optimize accesses when iterating over `len`. For example:https://github.com/golang/go/blob/4469f5446ad754d87afb340f59c213aec2dc4fb8/src/compress/flate/huffman_bit_writer.go#L211-L213

I would propose that instead of having in `huffmanEncoder` a slice of `hcode`, to have an `hcode` struct holding a slice of `code` and a slice of `len`. It would optimize the utilization of the CPU cache lines.

Here are the results I'm getting while comparing both benchmarks locally (x86) with `benstat`:

```
name                         old time/op    new time/op    delta
Decode/Digits/Huffman/1e4-4     112µs ± 0%     123µs ± 0%   ~     (p=1.000 n=1+1)
Decode/Digits/Huffman/1e5-4    1.07ms ± 0%    1.06ms ± 0%   ~     (p=1.000 n=1+1)
Decode/Digits/Huffman/1e6-4    10.7ms ± 0%    10.5ms ± 0%   ~     (p=1.000 n=1+1)
Decode/Newton/Huffman/1e4-4     131µs ± 0%     124µs ± 0%   ~     (p=1.000 n=1+1)
Decode/Newton/Huffman/1e5-4    1.21ms ± 0%    1.17ms ± 0%   ~     (p=1.000 n=1+1)
Decode/Newton/Huffman/1e6-4    11.5ms ± 0%    11.4ms ± 0%   ~     (p=1.000 n=1+1)
Encode/Digits/Huffman/1e4-4    45.5µs ± 0%    34.1µs ± 0%   ~     (p=1.000 n=1+1)
Encode/Digits/Huffman/1e5-4     445µs ± 0%     322µs ± 0%   ~     (p=1.000 n=1+1)
Encode/Digits/Huffman/1e6-4    4.32ms ± 0%    3.16ms ± 0%   ~     (p=1.000 n=1+1)
Encode/Newton/Huffman/1e4-4    59.6µs ± 0%    50.0µs ± 0%   ~     (p=1.000 n=1+1)
Encode/Newton/Huffman/1e5-4     507µs ± 0%     390µs ± 0%   ~     (p=1.000 n=1+1)
Encode/Newton/Huffman/1e6-4    5.00ms ± 0%    3.86ms ± 0%   ~     (p=1.000 n=1+1)

name                         old speed      new speed      delta
Decode/Digits/Huffman/1e4-4  89.6MB/s ± 0%  81.5MB/s ± 0%   ~     (p=1.000 n=1+1)
Decode/Digits/Huffman/1e5-4  93.2MB/s ± 0%  94.7MB/s ± 0%   ~     (p=1.000 n=1+1)
Decode/Digits/Huffman/1e6-4  93.4MB/s ± 0%  95.2MB/s ± 0%   ~     (p=1.000 n=1+1)
Decode/Newton/Huffman/1e4-4  76.5MB/s ± 0%  81.0MB/s ± 0%   ~     (p=1.000 n=1+1)
Decode/Newton/Huffman/1e5-4  82.9MB/s ± 0%  85.5MB/s ± 0%   ~     (p=1.000 n=1+1)
Decode/Newton/Huffman/1e6-4  86.7MB/s ± 0%  87.8MB/s ± 0%   ~     (p=1.000 n=1+1)
Encode/Digits/Huffman/1e4-4   220MB/s ± 0%   293MB/s ± 0%   ~     (p=1.000 n=1+1)
Encode/Digits/Huffman/1e5-4   225MB/s ± 0%   311MB/s ± 0%   ~     (p=1.000 n=1+1)
Encode/Digits/Huffman/1e6-4   232MB/s ± 0%   316MB/s ± 0%   ~     (p=1.000 n=1+1)
Encode/Newton/Huffman/1e4-4   168MB/s ± 0%   200MB/s ± 0%   ~     (p=1.000 n=1+1)
Encode/Newton/Huffman/1e5-4   197MB/s ± 0%   257MB/s ± 0%   ~     (p=1.000 n=1+1)
Encode/Newton/Huffman/1e6-4   200MB/s ± 0%   259MB/s ± 0%   ~     (p=1.000 n=1+1)

name                         old alloc/op   new alloc/op   delta
Decode/Digits/Huffman/1e4-4    40.5kB ± 0%    40.5kB ± 0%   ~     (all equal)
Decode/Digits/Huffman/1e5-4    40.5kB ± 0%    40.5kB ± 0%   ~     (all equal)
Decode/Digits/Huffman/1e6-4    40.6kB ± 0%    40.6kB ± 0%   ~     (all equal)
Decode/Newton/Huffman/1e4-4    41.2kB ± 0%    41.2kB ± 0%   ~     (all equal)
Decode/Newton/Huffman/1e5-4    45.0kB ± 0%    45.0kB ± 0%   ~     (p=1.000 n=1+1)
Decode/Newton/Huffman/1e6-4    79.9kB ± 0%    79.9kB ± 0%   ~     (all equal)
Encode/Digits/Huffman/1e4-4     0.00B          0.00B        ~     (all equal)
Encode/Digits/Huffman/1e5-4     1.00B ± 0%     1.00B ± 0%   ~     (all equal)
Encode/Digits/Huffman/1e6-4     16.0B ± 0%     12.0B ± 0%   ~     (p=1.000 n=1+1)
Encode/Newton/Huffman/1e4-4     0.00B          0.00B        ~     (all equal)
Encode/Newton/Huffman/1e5-4     1.00B ± 0%     1.00B ± 0%   ~     (all equal)
Encode/Newton/Huffman/1e6-4     19.0B ± 0%     14.0B ± 0%   ~     (p=1.000 n=1+1)

name                         old allocs/op  new allocs/op  delta
Decode/Digits/Huffman/1e4-4      5.00 ± 0%      5.00 ± 0%   ~     (all equal)
Decode/Digits/Huffman/1e5-4      5.00 ± 0%      5.00 ± 0%   ~     (all equal)
Decode/Digits/Huffman/1e6-4      5.00 ± 0%      5.00 ± 0%   ~     (all equal)
Decode/Newton/Huffman/1e4-4      14.0 ± 0%      14.0 ± 0%   ~     (all equal)
Decode/Newton/Huffman/1e5-4      23.0 ± 0%      23.0 ± 0%   ~     (all equal)
Decode/Newton/Huffman/1e6-4       161 ± 0%       161 ± 0%   ~     (all equal)
Encode/Digits/Huffman/1e4-4      0.00           0.00        ~     (all equal)
Encode/Digits/Huffman/1e5-4      0.00           0.00        ~     (all equal)
Encode/Digits/Huffman/1e6-4      0.00           0.00        ~     (all equal)
Encode/Newton/Huffman/1e4-4      0.00           0.00        ~     (all equal)
Encode/Newton/Huffman/1e5-4      0.00           0.00        ~     (all equal)
Encode/Newton/Huffman/1e6-4      0.00           0.00        ~     (all equal)
```